Fix ellipsize crash in demo app
The app crashes when trying to ellipsize long token/request codes. Seems to be a bug in the framework.
We can display the entire token/code and instead of changing the ConstraintLayout constraints we can simply use a LinearLayout. And since this is the only place where we use a ConstraintLayout we can get rid of the dependency. ConstraintLayout is a bit overkill here anyway.